ai_service/requirements.txt
# requirements file updated to version with more state data

#Python 3.11.1
# Will need to load graphvis to view graph diagrams
# linux: sudo apt-get install python3-dev graphviz libgraphviz-dev pkg-config
# mac: brew install graphviz
# see install instructions at https://pygraphviz.github.io/documentation/stable/install.html
# This is troublesome so pygraphvis is commented out below.

#Models Used:
# OpenAI:
#  - gpt-3.5, gpt-4o 
# tavily search tool

# also see pipe-freeze file - also without pygraphvis

python-dotenv~=1.0.1
tavily-python==0.3.3
duckduckgo_search==5.3.1b1 
beautifulsoup4==4.12.3
openai
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain>=0.3.0
langchain-text-splitters>=0.3.0
aiosqlite==0.20.0
pygments==2.17.2
# uncomment after installing, see above. Without this, graph display is off, but all else works
#pygraphviz==1.13
gradio==4.31.3
# added post-launch
httpx==0.27.0
fastapi==0.111.0
starlette==0.37.2
transformers
langchain-google-genai
faster_whisper
ipywidgets==8.1.2
jupyter==1.0.0
langchain_experimental>=0.0.47
tiktoken
weaviate-client
langchain-weaviate
langchain-huggingface
langchain-unstructured 
unstructured-client 
unstructured 
python-magic
google-genai